# Implement faster RSA key check