@bartonjs 2024-02-05
/backport to release/8.0-staging