@sebastienros 2024-02-12
FYI impact on aspnet startup time with https

![image](https://github.com/dotnet/runtime/assets/1165805/f9cc9a9c-7070-4e37-a947-23bd3eefccb1)