@adamsitnik 2024-02-22
@stephentoub this perf improvement is blog-post worthy!