@stephentoub 2024-02-22
> @stephentoub this perf improvement is blog-post worthy!

Not for a 9 post; this was backported to 8.